JieKou-Auto/requirements.txt
selenium>=4.15.0
pytest>=7.4.0
pyyaml>=6.0.1
allure-pytest>=2.13.2
webdriver-manager>=4.0.1
pytest-selenium>=4.0.0
requests==2.32.3
urllib3==2.2.1
allure-python-commons